MolecularMechanicsDict, SimulationDict and TrajectoryDict as native objects.
